Fix race condition where session disconnect pending a library logon reply can bump sequence index
